\name{data_symbolic}
\alias{data_symbolic}
\docType{data}
\title{Symbolic interval data}
\description{Artificially generated symbolic interval data}
\format{3-dimensional array: 125 objects, 6 variables, third dimension represents  begining and end of interval, 5-class structure}
\source{Artificially generated data}
\keyword{datasets}
