\name{SO2Simple}
\alias{SO2Simple}
\title{Change of representation of symbolic data from symbolic data table to simple form}
\description{Change of representation of symbolic data from symbolic data table to simple form}
\usage{
SO2Simple(sd)
}
\arguments{
\item{sd}{Symbolic data table in full form}
}
\details{
see \code{\link{symbolic.object}} for symbolic data table R structure representation
}
\value{
symbolic interval data: a 3-dimensional table, first dimension represents object number, second dimension - variable number, and third dimension contains lower- and upper-bounds of intervals
}
\author{
Andrzej Dudek \email{andrzej.dudek@ue.wroc.pl} 

Department of Econometrics and Computer Science, University of Economics, Wroclaw, Poland \url{http://keii.ue.wroc.pl/symbolicDA/}
}
\references{
Billard L., Diday E. (eds.) (2006), \emph{Symbolic Data Analysis, Conceptual Statistics and Data Mining}, John Wiley & Sons, Chichester.

Bock H.H., Diday E. (eds.) (2000), \emph{Analysis of symbolic data. Explanatory methods for extracting statistical information from complex data}, Springer-Verlag, Berlin.

Diday E., Noirhomme-Fraiture M. (eds.) (2008), \emph{Symbolic Data Analysis with SODAS Software}, John Wiley & Sons, Chichester.
}
\seealso{
\code{link{simple2SO}}
}
\examples{
# Example will be available in next version of package, thank You for your patience :-)
}
\keyword{cluster}
