\name{summary.symbolic}
\alias{summary.symbolic}
\title{Short description of summary data table}
\description{Short description of summary data table}
\usage{
\method{summary}{symbolic}(object,...)
}
\arguments{
\item{object}{Symbolic data table}
\item{...}{...}
}
\details{
Section details goes here
}
\value{
Section details goes here
}
\author{
Andrzej Dudek \email{andrzej.dudek@ue.wroc.pl} 

Department of Econometrics and Computer Science, University of Economics, Wroclaw, Poland \url{http://keii.ue.wroc.pl/symbolicDA}
}
\references{
Billard L., Diday E. (red.) (2006), \emph{Symbolic Data Analysis, Conceptual Statistics and Data Mining}, John Wiley & Sons, Chichester.

Bock H.H., Diday E. (eds.) (2000), \emph{Analysis of symbolic data. Explanatory methods for extracting statistical information from complex data}, Springer-Verlag, Berlin.

Diday E., Noirhomme-Fraiture M. (red.) (2008), \emph{Symbolic Data Analysis with SODAS Software}, John Wiley & Sons, Chichester.
}
\seealso{
\code{\link{symbolic.object}}
}
\examples{
#Example will be available in next version of package, thank You for your patience :-)
}
\keyword{cluster}